Move test related io out of utils.
